requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
-e git+https://github.com/misli/trml2pdf.git@registerFontFamily#egg=trml2pdf
backcall==0.1.0           # via ipython
certifi==2018.8.24        # via requests
chardet==3.0.4            # via requests
cmsplugin-filer==1.1.3
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid, social-auth-core
django-appconf==1.0.2     # via cmsplugin-filer
django-classy-tags==0.8.0  # via django-cms, django-sekizai
django-cms==3.5.2
django-countries==5.3.2
django-excel==0.0.10
django-filer==1.3.2
django-formtools==2.1     # via django-cms
django-ganalytics==0.7.0
django-haystack==2.7.0
django-localflavor==1.6.2
django-mptt==0.8.7        # via django-filer
django-polymorphic==2.0.3  # via django-filer
django-sekizai==0.10.0    # via cmsplugin-filer, django-cms
django-staticfiles-downloader==0.1.5
django-treebeard==4.3     # via django-cms
django-verified-email-field==1.4.2
django==1.11.15
djangocms-admin-style==1.2.8  # via django-cms
djangocms-attributes-field==0.3.0  # via cmsplugin-filer, djangocms-link, djangocms-style, djangocms-video
djangocms-googlemap==1.1.1
djangocms-link==2.1.2
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
easy-thumbnails==2.5      # via cmsplugin-filer, django-filer
html2rml==0.3.0
html5lib==0.9999999       # via djangocms-text-ckeditor
idna==2.7                 # via requests
ipython-genutils==0.2.0   # via traitlets
ipython==6.5.0
jedi==0.12.1              # via ipython
lml==0.0.4                # via pyexcel, pyexcel-io
mysqlclient==1.3.13
oauthlib==2.1.0           # via requests-oauthlib, social-auth-core
parso==0.3.1              # via jedi
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.2.0             # via djangocms-text-ckeditor, easy-thumbnails, reportlab
prompt-toolkit==1.0.15    # via ipython
psycopg2-binary==2.7.5
ptyprocess==0.6.0         # via pexpect
pyexcel-io==0.5.9.1       # via pyexcel, pyexcel-xlsxw
pyexcel-webio==0.1.4      # via django-excel
pyexcel-xlsxw==0.4.2
pyexcel==0.5.9.1          # via django-excel, pyexcel-webio
pygments==2.2.0           # via ipython
pyjwt==1.6.4              # via social-auth-core
pypdf2==1.26.0
python-memcached==1.59
python3-openid==3.1.0     # via social-auth-core
pytz==2018.5              # via django, django-staticfiles-downloader
raven==6.9.0
reportlab==3.5.6
requests-oauthlib==1.0.0  # via social-auth-core
requests==2.19.1          # via requests-oauthlib, social-auth-core
simplegeneric==0.8.1      # via ipython
six==1.11.0               # via html5lib, prompt-toolkit, python-memcached, social-auth-app-django, social-auth-core, traitlets
social-auth-app-django==2.1.0
social-auth-core==1.7.0   # via social-auth-app-django
texttable==1.4.0          # via pyexcel
traitlets==4.3.2          # via ipython
unidecode==0.4.21         # via django-filer
urllib3==1.23             # via requests
uwsgi==2.0.17.1
wcwidth==0.1.7            # via prompt-toolkit
xlsxwriter==0.9.3         # via pyexcel-xlsxw
